Undo changes to existing integration test package names
fix file paths
