+++ http/restful-ws-jakarta-integration-tests/pom.xml
@@ -22,7 +22,7 @@
         <module>restful-ws-jakarta-common</module>
         <module>restful-ws-microprofile</module>
         <module>restful-ws-resteasy</module>
-        <module>restful-ws-jakarta-jersey</module>
+        <module>restful-ws-jersey</module>
     </modules>
 
 </project>
