+++ http/restful-ws-jakarta-integration-tests/restful-ws-jersey/src/main/java/com/github/hanleyt/JerseyExtension.java
